Wormhole gun fixes and changes

+++ code/game/objects/effects/portals.dm
@@ -10,17 +10,19 @@
 	var/creator = null
 	anchored = 1.0
 	var/precision = 1 // how close to the portal you will teleport. 0 = on the portal, 1 = adjacent
+	var/can_multitool_to_remove = 0
 
 /obj/effect/portal/Bumped(mob/M as mob|obj)
 	src.teleport(M)
 
-/obj/effect/portal/New(loc, turf/target, creator)
+/obj/effect/portal/New(loc, turf/target, creator, removal_delay=300)
 	portals += src
 	src.loc = loc
 	src.target = target
 	src.creator = creator
-	spawn(300)
-		qdel(src)
+	if(removal_delay)
+		spawn(removal_delay)
+			qdel(src)
 	return
 
 /obj/effect/portal/Destroy()
@@ -48,3 +50,7 @@
 			do_teleport(M, locate(rand(5, world.maxx - 5), rand(5, world.maxy -5), 3), 0)
 		else
 			do_teleport(M, target, precision) ///You will appear adjacent to the beacon
+
+/obj/effect/portal/attackby(obj/item/A, mob/user)
+	if(istype(A, /obj/item/device/multitool) && can_multitool_to_remove)
+		qdel(src)
